requirements.txt
PyMuPDF
numpy
openai
requests
rank-bm25
scikit-learn
networkx
matplotlib
tqdm
Pillow
pytesseract
python-docx
faiss-cpu
pandas
ipywidgets
rdflib
pyvis
datasets
transformers
accelerate